Fix PR comments
- small css fixes 
- move ReservationFormItem component to its own file
